refactor: 당첨 번호를 관리하던 클래스를 로또 넘버 클래스로 관리하게끔 변경함 - 번호를 관리하는 일을 같이하고 있어, 코드가 중복되어 로또 번호 클래스로 관리함
