+++ src/main/java/lotto/view/InputView.java
@@ -1,12 +1,15 @@
 package lotto.view;
 
+import java.util.List;
 import java.util.Scanner;
+import java.util.stream.Collectors;
 
 import static java.util.Arrays.stream;
 
 public class InputView {
-    private static final String INVALID_WINNING_NUMBER = "당첨 번호를 정확하게 입력해주세요.";
+    private static final String INVALID_WINNING_NUMBER = "로또 번호를 정확하게 입력해주세요.";
     private static final Scanner SCAN = new Scanner(System.in);
+    private static final String SEPARATOR_COMMA = ",";
 
     private InputView() {
     }
@@ -15,15 +18,17 @@ public class InputView {
         return Integer.parseInt(SCAN.nextLine());
     }
 
-    public static String[] inputWinningNumbers() {
+    public static List<Integer> inputWinningNumbers() {
         return splitWinningNumbers(SCAN.nextLine());
     }
 
-    private static String[] splitWinningNumbers(String inputWinningNumbers) {
+    private static List<Integer> splitWinningNumbers(String inputWinningNumbers) {
         validateNullOrEmpty(inputWinningNumbers);
-        return stream(inputWinningNumbers.split(","))
+        return stream(inputWinningNumbers.split(SEPARATOR_COMMA))
                 .map(String::trim)
-                .toArray(String[]::new);
+                .mapToInt(Integer::parseInt)
+                .boxed()
+                .collect(Collectors.toList());
     }
 
     private static void validateNullOrEmpty(String inputWinningNumbers) {
